MasgitoffTest: Move code to MasgitoffIo class
