+++ src/main/java/org/mastodon/mamut/collaboration/io/ModelSerializer.java
